# project/backend/document/default_OpenAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "个人贷款管理系统 API",
    "description": "个人贷款管理系统后端API文档",
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/orders/{orderId}/repay": {
      "post": {
        "tags": [
          "订单管理"
        ],
        "summary": "还款操作",
        "description": "用户对指定贷款订单进行还款操作",
        "operationId": "repayOrder",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "订单ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultUserGetOrderResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/approval/check": {
      "post": {
        "tags": [
          "审批管理"
        ],
        "summary": "审批操作",
        "description": "管理员对贷款申请进行审批操作（批准/拒绝）",
        "operationId": "manualCheck",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultManualCheckResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/orders/{orderId}": {
      "get": {
        "tags": [
          "订单管理"
        ],
        "summary": "获取订单详情",
        "description": "用户获取指定贷款订单的详细信息",
        "operationId": "getUserOrder",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "订单ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultUserGetOrderResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/approval/detail/{loanApplicationId}": {
      "get": {
        "tags": [
          "审批管理"
        ],
        "summary": "获取申请详情",
        "description": "管理员获取贷款申请的详细信息",
        "operationId": "getApplicationDetail",
        "parameters": [
          {
            "name": "loanApplicationId",
            "in": "path",
            "description": "申请ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultApplicationDetailResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResultUserGetOrderResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/UserGetOrderResponse"
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          }
        },
        "description": "API响应结果"
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "userId": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "productId": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "status": {
            "type": "string",
            "enum": [
              "正常",
              "已逾期",
              "已完成"
            ],
            "refType": null
          },
          "repaidAmount": {
            "type": "number",
            "refType": null
          },
          "loanAmount": {
            "type": "number",
            "refType": null
          },
          "interestRate": {
            "type": "number",
            "refType": null
          },
          "repaidType": {
            "type": "string",
            "enum": [
              "等额本息",
              "等额本金",
              "先息后本",
              "一次性还本付息"
            ],
            "refType": null
          },
          "loanPeriod": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "term": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "currentTerm": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "contract": {
            "type": "string",
            "refType": null
          },
          "overdueDays": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "startTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          }
        }
      },
      "UserGetOrderResponse": {
        "type": "object",
        "properties": {
          "productName": {
            "type": "string"
          },
          "order": {
            "$ref": "#/components/schemas/Order"
          }
        },
        "description": "响应数据"
      },
...
      "ManualCheckRequest": {
        "type": "object",
        "properties": {
          "loanApplicationId": {
            "type": "integer",
            "format": "int64"
          },
          "approved": {
            "type": "boolean"
          },
          "manualRejectReason": {
            "type": "string"
          }
        },
        "description": "审批请求"
      },
      "ApiResultManualCheckResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/ManualCheckResponse"
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          }
        },
        "description": "API响应结果"
      },
      "ManualCheckResponse": {
        "type": "object",
        "properties": {
          "loanApplicationId": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string",
            "enum": [
              "审核中",
              "已通过",
              "AI拒绝",
              "人工拒绝",
              "已取消"
            ]
          },
          "rejectReason": {
            "type": "string"
          },
          "reviewTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "description": "响应数据"
      },
...
      "LoanApplication": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "userId": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "productId": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "status": {
            "type": "string",
            "enum": [
              "审核中",
              "已通过",
              "AI拒绝",
              "人工拒绝",
              "已取消"
            ],
            "refType": null
          },
          "loanAmount": {
            "type": "number",
            "refType": null
          },
          "interestRate": {
            "type": "number",
            "refType": null
          },
          "loanPeriod": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "term": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "repaidType": {
            "type": "string",
            "enum": [
              "等额本息",
              "等额本金",
              "先息后本",
              "一次性还本付息"
            ],
            "refType": null
          },
          "rejectReason": {
            "type": "string",
            "refType": null
          },
          "applyTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "reviewTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          }
        }
      },
      "User": {
        "required": [
          "password",
          "phone"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "userName": {
            "maxLength": 20,
            "minLength": 2,
            "type": "string",
            "refType": null
          },
          "avatar": {
            "type": "string",
            "refType": null
          },
          "phone": {
            "pattern": "^1[3-9]\\d{9}$",
            "type": "string",
            "refType": null
          },
          "password": {
            "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[!@#$%&*()_+=])[A-Za-z\\d!@#$%&*()_+=]{8,20}$",
            "type": "string",
            "refType": null
          },
          "role": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          }
        }
      },
      "UserCert": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "idCard": {
            "type": "string",
            "refType": null
          },
          "creditScore": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "bankCardId": {
            "type": "string",
            "refType": null
          },
          "workCertId": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "triCertId": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "immovableCertId": {
            "type": "integer",
            "format": "int32",
            "refType": null
          }
        }
      },
...
      "ApiResultApplicationDetailResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/ApplicationDetailResponse"
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          }
        },
        "description": "API响应结果"
      },
      "ApplicationDetailResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "userCert": {
            "$ref": "#/components/schemas/UserCert"
          },
          "application": {
            "$ref": "#/components/schemas/LoanApplication"
          }
        },
        "description": "响应数据"
      }
    }
  }
}